plugins/yls-yara/pyproject.toml
[tool.poetry]
name = "yls-yara"
version = "0.1.0"
description = "YLS plugin adding linting using yara-python."
authors = ["Matej Kastak <matej.kastak@avast.com>"]
maintainers = ["Matej Kašťák <matej.kastak@avast.com>"]
readme = "README.md"
license = "MIT"
include = ["py.typed", "LICENSE"]

[tool.poetry.dependencies]
python = "^3.8"
yara-python = "^4.2.0"
yls = "^1.0.0"

[tool.poetry.dev-dependencies]
pytest = "^7.1.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.plugins.yls]
yara = "yls_yara"
